refactor: change components folder structur
