+++ src/app/NotificationsProvider.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useRecoilValue } from 'recoil';
 
 import { appNotificationState } from '@/atoms/RecoilState';
-import AppNotifications from '@/components/feedback/AppNotifications';
+import AppNotifications from '@/components/ui/AppNotifications';
 
 const NotificationsProvider: React.FC = () => {
   const appNotifications = useRecoilValue(appNotificationState);
