+++ src/app/admin/AdminEventList.tsx
@@ -6,9 +6,9 @@ import Link from 'next/link';
 import useTranslation from 'next-translate/useTranslation';
 import { useState } from 'react';
 
-import { createColumnHelper, DataTable } from '@/components/content';
 import EventEmailer from '@/components/event/EventEmailer';
 import { Drawer } from '@/components/layout';
+import DataTable, { createColumnHelper } from '@/components/ui/DataTable';
 const columnHelper = createColumnHelper<EventDto>();
 interface AdminEventListProps {
   eventinfo: EventDto[];
