+++ src/app/admin/page.tsx
@@ -3,11 +3,11 @@
 import { UserDto } from '@losol/eventuras/models/UserDto';
 import useTranslation from 'next-translate/useTranslation';
 
-import { Heading } from '@/components/content';
-import { Loading } from '@/components/feedback';
-import { InputAutoComplete } from '@/components/inputs/Input';
-import Link from '@/components/inputs/Link';
+import { InputAutoComplete } from '@/components/forms/Input';
 import { Container, Layout } from '@/components/layout';
+import Heading from '@/components/ui/Heading';
+import Link from '@/components/ui/Link';
+import Loading from '@/components/ui/Loading';
 import { useEvents } from '@/hooks/apiHooks';
 import { getUsers } from '@/utils/api/functions/users';
 import Environment from '@/utils/Environment';
