+++ src/app/user/events/[id]/registration/components/RegistrationCustomize.tsx
@@ -1,9 +1,9 @@
 import useTranslation from 'next-translate/useTranslation';
 import { useForm } from 'react-hook-form';
 
-import { Heading } from '@/components/content';
-import { Button } from '@/components/inputs';
-import Checkbox from '@/components/inputs/Checkbox';
+import Checkbox from '@/components/forms/Checkbox';
+import Button from '@/components/ui/Button';
+import Heading from '@/components/ui/Heading';
 import { RegistrationProduct } from '@/types/RegistrationProduct';
 
 type SubmitCallback = (values: Map<string, number>) => void;
