+++ src/app/user/events/[id]/registration/components/RegistrationPayment.tsx
@@ -3,10 +3,10 @@ import useTranslation from 'next-translate/useTranslation';
 import { useEffect, useState } from 'react';
 import { SubmitHandler, useForm } from 'react-hook-form';
 
-import { Heading } from '@/components/content';
-import { Button } from '@/components/inputs';
-import { defaultInputStyle, InputText } from '@/components/inputs/Input';
+import { defaultInputStyle, InputText } from '@/components/forms/Input';
 import Container from '@/components/layout/Container';
+import Button from '@/components/ui/Button';
+import Heading from '@/components/ui/Heading';
 import PaymentFormValues from '@/types/PaymentFormValues';
 import { UserProfile } from '@/types/UserProfile';
