+++ src/atoms/RecoilState.ts
@@ -1,6 +1,6 @@
 import { atom } from 'recoil';
 
-import { AppNotificationOptions } from '@/components/feedback/AppNotifications';
+import { AppNotificationOptions } from '@/components/ui/AppNotifications';
 
 export const appNotificationState = atom<AppNotificationOptions[]>({
   key: 'appNotificationState',
