--- src/components/content/index.ts
@@ -1,5 +1,0 @@
-export { default as Text } from '../content/Text';
-export { default as Card } from './Card';
-export { createColumnHelper, default as DataTable } from './DataTable';
-export { default as Heading } from './Heading';
-export { default as Image } from './Image';
