--- src/components/feedback/index.ts
@@ -1,3 +1,0 @@
-export { default as Loading } from './Loading';
-export { default as Spinner } from './Spinner';
-export { default as Unauthorized } from './Unauthorized';
