--- src/components/inputs/index.ts
@@ -1,2 +1,0 @@
-export { default as Button } from './Button';
-export { default as Switch } from './Switch';
