+++ src/components/layout/Drawer.tsx
@@ -1,9 +1,9 @@
 import { IconX } from '@tabler/icons-react';
 import React, { HTMLAttributes, ReactNode } from 'react';
 
-import { Heading } from '@/components/content';
-import { Button } from '@/components/inputs';
 import { Portal } from '@/components/layout';
+import Button from '@/components/ui/Button';
+import Heading from '@/components/ui/Heading';
 
 export interface DrawerProps {
   isOpen: boolean;
