+++ src/components/ui/Image.cy.tsx
