+++ src/components/ui/Image.stories.tsx
